backend/requirements.txt
chromadb==0.5.5
Flask==2.3.2
Flask_Cors==4.0.1
openai==0.27.4
python-dotenv==1.0.1
python_docx==1.1.2
